Don't send loading messages to pager
When just stdout of the Console object is set, it sends both what would
normally go to stdout and stderr to that one destination instead, so
even the "Fetching log messages" loading message gets sent to the stdin
of the pager (less) and get displayed as a result.

The message should only go to the parent process's stdout.