stop ignoring top level errors
It's best we can see errors when they happen.

The main thing we were working around was piping output from this
command to a pager in the shell (invoking us as `k-kdesvn ... | less`),
which is no longer needed now that we integrate piping into less in the
log command already.